requirements.txt
Flask
flask_cors
pymongo
pymongo[srv]
python-dotenv